Add options to lak list whatif
Added three new options to lak list whatif that allows user to print
long name (or not), short name and the quantity of an asset
that need to be bought or sold. The quantity is only printed for asset
that support it (e.g. TickerAssets) and is approximate (whole) number of
shares.